requirements.txt
OpenSSL
logging
pytz
ndg.httpsclient
pyasn1
datetime